+++ extension/src/Settings.ts
@@ -17,6 +17,7 @@ export enum Setting {
     XlfHighlightsDecoration,
     UseExternalTranslationTool,
     UseDTS,
+    DTSProjectId,
     ReplaceSelfClosingXlfTags,
     SearchOnlyXlfFiles,
     MatchTranslation,
@@ -64,6 +65,7 @@ export class Settings {
         this.SettingCollection[Setting.XlfHighlightsDecoration] = this.config.get('XlfHighlightsDecoration');
         this.SettingCollection[Setting.UseExternalTranslationTool] = this.config.get('UseExternalTranslationTool');
         this.SettingCollection[Setting.UseDTS] = this.config.get('UseDTS');
+        this.SettingCollection[Setting.DTSProjectId] = this.config.get('DTS ProjectId');
         this.SettingCollection[Setting.ReplaceSelfClosingXlfTags] = this.config.get('ReplaceSelfClosingXlfTags');
         this.SettingCollection[Setting.SearchOnlyXlfFiles] = this.config.get('SearchOnlyXlfFiles');
         this.SettingCollection[Setting.MatchTranslation] = this.config.get('MatchTranslation');