+++ extension/src/extension.ts
@@ -18,6 +18,7 @@ export function activate(context: vscode.ExtensionContext) {
         let commandlist = [
                 vscode.commands.registerCommand('nab.RefreshXlfFilesFromGXlf', () => { NABfunctions.refreshXlfFilesFromGXlf(); }),
                 vscode.commands.registerCommand('nab.FormatCurrentXlfFileForDTS', () => { NABfunctions.formatCurrentXlfFileForDts(); }),
+                vscode.commands.registerCommand('nab.OpenDTS', () => { NABfunctions.openDTS(); }),
                 vscode.commands.registerCommand('nab.FindNextUnTranslatedText', () => { NABfunctions.findNextUnTranslatedText(); }),
                 vscode.commands.registerCommand('nab.SetTranslationUnitToTranslated', () => { NABfunctions.setTranslationUnitToTranslated(); }),
                 vscode.commands.registerCommand('nab.SetTranslationUnitToSignedOff', () => { NABfunctions.setTranslationUnitToSignedOff(); }),